break circdep in output.ts
